Began refactoring of tests to eliminate duplicated code
